+++ src/core/messaging/OnNewMessageOptionalParams.ts
@@ -1,3 +1,5 @@
 export default interface OnNewMessageOptionalParams {
-    rehydrate: boolean;
+    rehydrate?: boolean;
+    pollingInterval?: number;
+    disablePolling?: boolean;
 }